+++ src/main/java/com/podnet/podnet/entity/User.java
@@ -11,21 +11,61 @@ import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
-
-import jakarta.persistence.*;
-import lombok.Data;
-import java.util.List;
-
-@Entity
 @Data
-public class User {
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "users_table")
+public class User implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
     private Long id;
 
+    @Column(name = "username")
     private String username;
+
+    @Column(name = "email")
+    private String email;
+
+    @Column(name = "password")
+    private String password;
+
+    @Enumerated(EnumType.STRING)
+    private Role role;
+
+    @OneToMany(mappedBy = "user")
+    private List<Token> tokens;
 
     @OneToMany(mappedBy = "author")
     private List<Message> messages;
-}
+//
+//    @ManyToMany(mappedBy = "users")
+//    private List<Chat> chats;
 
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority((role.name())));
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+}
