--- src/main/java/com/podnet/podnet/entity/UserAUth.java
@@ -1,73 +1,0 @@
-package com.podnet.podnet.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.Collection;
-import java.util.List;
-
-public class UserAUth {
-    @Data
-    @AllArgsConstructor
-    @NoArgsConstructor
-    @Entity
-    @Table(name = "users_table")
-    public static class UserAuth implements UserDetails {
-        @Id
-        @GeneratedValue(strategy = GenerationType.IDENTITY)
-        @Column(name = "id")
-        private Long id;
-
-        @Column(name = "username")
-        private String username;
-
-        @Column(name = "email")
-        private String email;
-
-        @Column(name = "password")
-        private String password;
-
-        @Enumerated(EnumType.STRING)
-        private Role role;
-
-        @OneToMany(mappedBy = "user")
-        private List<Token> tokens;
-
-        @OneToMany(mappedBy = "author")
-        private List<Message> messages;
-
-        @ManyToMany(mappedBy = "users")
-        private List<Chat> chats;
-
-
-        @Override
-        public Collection<? extends GrantedAuthority> getAuthorities() {
-            return List.of(new SimpleGrantedAuthority((role.name())));
-        }
-
-        @Override
-        public boolean isAccountNonExpired() {
-            return true;
-        }
-
-        @Override
-        public boolean isAccountNonLocked() {
-            return true;
-        }
-
-        @Override
-        public boolean isCredentialsNonExpired() {
-            return true;
-        }
-
-        @Override
-        public boolean isEnabled() {
-            return true;
-        }
-    }
-}
