+++ src/main/java/com/podnet/podnet/service/UserServiceImpl.java
@@ -1,67 +1,72 @@
-//package com.podnet.podnet.service;
-//
-//import com.podnet.podnet.entity.User;
-//import com.podnet.podnet.repository.UserRepository;
-//import com.podnet.podnet.service.UserService;
-//import org.springframework.security.core.Authentication;
-//import org.springframework.security.core.context.SecurityContextHolder;
-//import org.springframework.security.core.userdetails.UserDetails;
-//import org.springframework.security.core.userdetails.UsernameNotFoundException;
-//import org.springframework.stereotype.Service;
-//
-//import java.util.Optional;
-//
-//
-///**
-// * Сервис для работы с пользователями.
-// */
-//
-//@Service
-//public class UserServiceImpl implements UserService {
-//
-//    private final UserRepository userRepository;
-//
-//    public UserServiceImpl(UserRepository userRepository) {
-//        this.userRepository = userRepository;
-//    }
-//
-//
-//
-//    /**
-//     * Метод загружает пользователя по его имени.
-//     * Если пользователь не найден, выбрасывает исключение UsernameNotFoundException.
-//     *
-//     * @param username имя пользователя для поиска
-//     * @return найденный пользователь
-//     * @throws UsernameNotFoundException если пользователь не найден
-//     */
-////    @Override
-////    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-////        // Поиск пользователя в репозитории
-////        return userRepository.findByUsername(username)
-////                // Если пользователь не найден, выбрасываем исключение
-////                .orElseThrow(() -> new UsernameNotFoundException("Пользователь с именем " + username + " не найден"));
-////    }
-//
+package com.podnet.podnet.service;
+
+import com.podnet.podnet.entity.User;
+import com.podnet.podnet.repository.UserRepository;
+import com.podnet.podnet.service.UserService;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+
+/**
+ * Сервис для работы с пользователями.
+ */
+
+@Service
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+
+
+    /**
+     * Метод загружает пользователя по его имени.
+     * Если пользователь не найден, выбрасывает исключение UsernameNotFoundException.
+     *
+     * @param username имя пользователя для поиска
+     * @return найденный пользователь
+     * @throws UsernameNotFoundException если пользователь не найден
+     */
 //    @Override
-//    public boolean existsByUsername(String username) {
-//        User user = userRepository.findByUsername(username).orElse(null);
-//        if (user != null) {
-//            return true;
-//        }
-//        return false;
+//    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+//        // Поиск пользователя в репозитории
+//        return userRepository.findByUsername(username)
+//                // Если пользователь не найден, выбрасываем исключение
+//                .orElseThrow(() -> new UsernameNotFoundException("Пользователь с именем " + username + " не найден"));
 //    }
-//
-//    @Override
-//    public boolean existsByEmail(String email) {
-//        User user = userRepository.findByEmail(email).orElse(null);
-//        if (user != null) {
-//            return true;
-//        }
-//        return false;
-//    }
-//    public Optional<User> getCurrentUser() {
-//        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-//        return userRepository.findByUsername(auth.getName());
-//    }
-//}
+
+    @Override
+    public boolean existsByUsername(String username) {
+        User user = userRepository.findByUsername(username).orElse(null);
+        if (user != null) {
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public boolean existsByEmail(String email) {
+        User user = userRepository.findByEmail(email).orElse(null);
+        if (user != null) {
+            return true;
+        }
+        return false;
+    }
+    public Optional<User> getCurrentUser() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        return userRepository.findByUsername(auth.getName());
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+        return null;
+    }
+}
