+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.datasource.username=postgres
 spring.datasource.password=postgres
 spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
 spring.datasource.driver-class-name=org.postgresql.Driver
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=update
 
 server.port=8080
